Break out unit tests for packettype conversions and positioninfo
